requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
streamlit==1.28.0
qdrant-client==1.15.1
supabase>=2.3.0
openai>=1.50.0
python-dotenv==1.0.0
pysqlite3-binary
numpy>=1.24.0
pandas>=2.1.2
duckdb
PyPDF2
python-docx
torch==2.3.1
sentence-transformers>=2.2.0
requests>=2.31.0
pytz
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0